pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "roughbench"
version = "0.1.0"
description = "Opinionated benchmark for real-world LLM competence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "datasets>=3.0.0",
  "openai>=1.30.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
anthropic = [
  "anthropic>=0.40.0",
]

[project.scripts]
roughbench = "roughbench.cli:main"

[tool.setuptools.packages.find]
include = ["roughbench*"]
